refactor(fix): change package, class name, and visibility
