+++ core/src/main/java/com/scalar/db/storage/jdbc/RdbEngineStrategy.java
@@ -2,6 +2,6 @@ package com.scalar.db.storage.jdbc;
 
 import com.scalar.db.io.DataType;
 
-public abstract class RdbEngineStrategy {
-   protected abstract String getDataTypeForEngine(DataType dataType);
+abstract class RdbEngineStrategy {
+   abstract String getDataTypeForEngine(DataType dataType);
 }
